requirements.txt
grappelli
django-resized